Split lib.rs into multiple files
