+++ src/ffi_types/node_ref.rs
@@ -1,0 +1,3 @@
+use internal;
+
+pub type NodeRef = internal::YGNodeRef;
