+++ src/ffi_types/undefined.rs
@@ -1,0 +1,1 @@
+pub use std::f32::NAN as Undefined;
